requirements.txt
matplotlib==3.7.2
matplotlib==3.7.1
pandas==1.5.3
pandas==2.0.2
Pillow==10.0.0
Pillow==9.5.0
Pillow==10.0.1
torch==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
